requirements.txt
Flask
google-cloud-aiplatform
requests